Lots of changes have been made.
Added routing to the project, /visual is pretty much done. There is a
table in it as well as the map.

Refractoring the map component... still have the old one but it needs
to be deleted.

On the server side added a webscaper and tried to use the google api to
download pdf and do pdf parsing but couldn't get that to work. Need to
just fill in the data automatically into a .csv and parse it into a
json.